Correction du calcul de l'arrondi du niveau dans status.cpp
La méthode d'ajout du niveau à la statistique `str` utilisait une division entière avec `ceil(level / 10)`, ce qui pouvait donner des résultats incorrects pour certains niveaux. Désormais, la division se fait en flottant avec `ceil(level / 10.0)`, puis le résultat est converti en entier, assurant un arrondi correct même si le niveau n'est pas un multiple de 10. Cela améliore la précision du calcul de l'attaque de base.

+++ src/map/status.cpp
@@ -2765,7 +2765,7 @@ unsigned short status_base_atk(const struct block_list *bl, const struct status_
 			dstr = str / 10;
 			str += dstr*dstr;
 			str += dex / 5 + status->luk / 5;
-			str += ceil(level / 10); // [Stingor]
+			str += static_cast<int>(ceil(level / 10.0)); // [Stingor]
 #endif
 			break;
 		default:// Others
